requirements.txt
Django==4.2.7
djangorestframework==4.1.0
django-cors-headers==4.3.0
python-dotenv==1.0.1
groq
pillow
google
cloudunary